#30: Change to fluid layout with max-width on content (WIP), better way to import <Grid/>, reorganise <Input/> and <Textarea/> into root component folder
