+++ admin_ui/components/Grid/index.js
@@ -1,0 +1,1 @@
+export { default } from './Grid';
